``` feat(phase7): align RLM recursion with reference implementation and tighten persistence contracts
- Update daytona to 0.184.0 and fastapi to 0.136.3
- Add Phase 7 documentation covering history management, bounded child snapshots, token-aware compaction, explicit depth tracking, and fallback behavior
- Tighten PersistenceDep type from Any to PersistenceProtocol and raise RuntimeError when backend not initialized
- Remove ServerState compatibility proxy attributes (_SERVER_STATE_PROXY_ATTRS)
- Move RuntimeEventContext from api/events/
