+++ pyproject.toml
@@ -34,7 +34,7 @@ dependencies = [
     "dspy[optuna]==3.2.1",
 
     # Core orchestration and CLI/runtime ergonomics.
-    "daytona>=0.181.0,<1",
+    "daytona>=0.184.0,<1",
     "hydra-core>=1.3,<2",
     "prompt-toolkit>=3.0.50,<4",
     "rich>=14.3.3,<15",
@@ -56,7 +56,7 @@ dependencies = [
     "tomli>=2.0.0; python_version < '3.11'",
 
     # Web API surface.
-    "fastapi[standard]==0.136.1",
+    "fastapi[standard]==0.136.3",
     "joserfc>=1.0.1",
     "uvicorn>=0.47.0,<1",
