+++ src/fleet_rlm/api/events/__init__.py
@@ -14,6 +14,8 @@ from .events import (
     sanitize_event_payload,
     summarize_code_for_event,
 )
+from .project_chat import project_chat
+from .project_graph import project_graph
 from .step_builder import ExecutionStepBuilder
 
 __all__ = [
@@ -28,6 +30,8 @@ __all__ = [
     "ExecutionStepType",
     "ExecutionSubscription",
     "RuntimeEventContext",
+    "project_chat",
+    "project_graph",
     "sanitize_event_payload",
     "summarize_code_for_event",
 ]
