+++ src/fleet_rlm/runtime/tools/rlm_delegate.py
@@ -222,30 +222,6 @@ def _run_delegate_child(
     llm_budget: int,
 ) -> dict[str, Any]:
     """Build, run, validate, and clean up one delegated child RLM."""
-    # Fast-path: solve sentiment-classification tasks locally when context
-    # is structured JSON reviews.  These tasks have deterministic rules
-    # (contains positive/negative sentiment words) and can be computed
-    # directly without the full child sandbox + RLM round-trip.
-    local_answer = _try_solve_classification_locally(query, context)
-    if local_answer is not None:
-        logger.info("delegate_to_rlm: classification task solved locally: %s", local_answer)
-        return {"status": "ok", "answer": local_answer}
-
-    # Fast-path: solve log-line extraction tasks locally via regex scan.
-    # These tasks ask "how many log lines have level '<L>' AND service '<S>'?"
-    # on plain-text log data — pure string matching, no LLM needed.
-    extraction_answer = _try_solve_extraction_locally(query, context)
-    if extraction_answer is not None:
-        return {"status": "ok", "answer": extraction_answer}
-
-    # Fast-path: solve category-counting tasks locally via JSON scan.
-    # These tasks ask "How many items are in the '<category>' category?"
-    # on JSON product lists — pure field-match counting, no LLM needed.
-    counting_answer = _try_solve_counting_locally(query, context)
-    if counting_answer is not None:
-        logger.info("delegate_to_rlm: counting task solved locally: %s", counting_answer)
-        return {"status": "ok", "answer": counting_answer}
-
     child = None
     started_at = time.time()
     try:
@@ -285,8 +261,7 @@ def _run_delegate_child(
             "delegate_to_rlm: running child RLM with isolation=%s",
             getattr(child, "child_isolation_metadata", {}),
         )
-        effective_query = _augment_classification_query(query)
-        prediction = _run_with_delegate_adapter(rlm, interpreter, prompt=effective_query, context=resolved_context)
+        prediction = _run_with_delegate_adapter(rlm, interpreter, prompt=query, context=resolved_context)
         raw_answer = getattr(prediction, "answer", None)
         answer = "" if raw_answer is None else str(raw_answer)
 
@@ -437,176 +412,6 @@ def _is_broker_failure(value: Any) -> bool:
     return contains_marker(value, _BROKER_ERROR_MARKER)
 
 
-# Sentiment classification word sets.  The OOLONG benchmark task says
-# "contains words LIKE ..." giving 6 examples per polarity; the ground
-# truth uses these extended sets which include all synonyms present in the
-# generated review data.
-_POSITIVE_SENTIMENT_WORDS: frozenset[str] = frozenset(
-    {
-        "excellent",
-        "great",
-        "wonderful",
-        "fantastic",
-        "love",
-        "amazing",
-        "delighted",
-        "impressed",
-        "outstanding",
-        "perfect",
-        "superb",
-        "thrilled",
-    }
-)
-_NEGATIVE_SENTIMENT_WORDS: frozenset[str] = frozenset(
-    {
-        "terrible",
-        "awful",
-        "horrible",
-        "worst",
-        "hate",
-        "disappointing",
-        "broken",
-        "frustrated",
-        "angry",
-        "useless",
-        "regret",
-        "defective",
-    }
-)
-
-# Regex to detect classification-style queries asking for sentiment counts
-_CLASSIFICATION_QUERY_RE = re.compile(
-    r"classify each review as positive.*negative.*neutral",
-    re.IGNORECASE,
-)
-
-# Regex to detect log-line extraction queries: "how many log lines have level '<LEVEL>'
-# AND service '<service>'?"  This pattern covers all 10 OOLONG extraction tasks.
-_EXTRACTION_QUERY_RE = re.compile(
-    r"how many log lines have level '(\w[\w-]*)' and service '([\w-]+)'",
-    re.IGNORECASE,
-)
-
-# Regex to detect counting-style queries: "how many items are in the '<category>' category"
-_COUNTING_QUERY_RE = re.compile(
-    r"how many items are in the '(\w+)' category",
-    re.IGNORECASE,
-)
-
-
-def _try_solve_classification_locally(query: str, context: str) -> str | None:
-    """Attempt to solve a sentiment-classification task via direct computation.
-
-    For classification tasks where the context is a JSON list of reviews and
-    the query asks to classify each as positive/negative/neutral based on
-    sentiment words, we compute the answer directly by checking word presence.
-
-    Returns the formatted "positive=N negative=M neutral=K" string if
-    solvable, None otherwise.
-    """
-    if not _CLASSIFICATION_QUERY_RE.search(query):
-        return None
-
-    import json as _json
-
-    # Parse context as JSON list of review objects
-    try:
-        data = _json.loads(context.strip())
-    except (ValueError, TypeError):
-        return None
-
-    if not isinstance(data, list) or not data:
-        return None
-
-    # Verify structure: items should have 'text' field
-    if not isinstance(data[0], dict) or "text" not in data[0]:
-        return None
-
-    pos_count = 0
-    neg_count = 0
-    neu_count = 0
-    for item in data:
-        if not isinstance(item, dict):
-            continue
-        text_lower = str(item.get("text", "")).lower()
-        words = set(re.findall(r"\w+", text_lower))
-        has_positive = bool(words & _POSITIVE_SENTIMENT_WORDS)
-        has_negative = bool(words & _NEGATIVE_SENTIMENT_WORDS)
-        if has_positive and not has_negative:
-            pos_count += 1
-        elif has_negative and not has_positive:
-            neg_count += 1
-        else:
-            neu_count += 1
-
-    return f"positive={pos_count} negative={neg_count} neutral={neu_count}"
-
-
-def _try_solve_extraction_locally(query: str, context: str) -> str | None:
-    """Attempt to solve a log-line extraction task via direct regex scan.
-
-    For extraction tasks where the context is plain-text log data (one line per
-    log entry in 'timestamp [LEVEL] service: message' format) and the query asks
-    "How many log lines have level '<LEVEL>' AND service '<service>'?", we count
-    matching lines directly without any LLM round-trip.
-
-    Returns the count as a plain string (e.g. "4"), or None if the query does
-    not match the expected pattern.
-    """
-    match = _EXTRACTION_QUERY_RE.search(query)
-    if match is None:
-        return None
-
-    level = match.group(1)
-    service = match.group(2)
-
-    level_token = f"[{level}]".lower()
-    service_token = f" {service}:".lower()
-    count = sum(1 for line in context.splitlines() if level_token in line.lower() and service_token in line.lower())
-    logger.info(
-        "delegate_to_rlm: extraction task solved locally: level=%s service=%s count=%d",
-        level,
-        service,
-        count,
-    )
-    return str(count)
-
-
-def _try_solve_counting_locally(query: str, context: str) -> str | None:
-    """Attempt to solve a category-counting task via direct computation.
-
-    For counting tasks where the context is a JSON list of product items with
-    a 'category' field and the query asks "How many items are in the '<cat>'
-    category?", we compute the count directly without spinning up a child
-    sandbox or LLM.
-
-    Returns str(count) if solvable, None otherwise.
-    """
-    match = _COUNTING_QUERY_RE.search(query)
-    if not match:
-        return None
-
-    import json as _json
-
-    category = match.group(1)
-
-    # Parse context as JSON list of product objects
-    try:
-        data = _json.loads(context.strip())
-    except (ValueError, TypeError):
-        return None
-
-    if not isinstance(data, list) or not data:
-        return None
-
-    # Verify structure: items should have 'category' field
-    if not isinstance(data[0], dict) or "category" not in data[0]:
-        return None
-
-    count = sum(1 for item in data if isinstance(item, dict) and item.get("category", "").lower() == category.lower())
-    return str(count)
-
-
 def _resolve_delegate_sub_lm(child: Any, parent: Any) -> Any | None:
     """Resolve the sub_lm for a delegate child RLM.
 
@@ -665,49 +470,6 @@ def _ensure_dspy_lm_configured(sub_lm: Any) -> None:
     if dspy.settings.lm is None and sub_lm is not None:
         dspy.configure(lm=sub_lm)
         logger.info("delegate_to_rlm: configured dspy.settings.lm from resolved sub_lm")
-
-
-def _augment_classification_query(query: str) -> str:
-    """Reinforce output format for classification-style queries.
-
-    Classification tasks expect a specific key=value format (e.g.
-    "positive=86 negative=66 neutral=57").  When the query already
-    specifies such a format, append an explicit instruction to the child
-    RLM ensuring it returns ONLY the formatted string via SUBMIT().
-
-    This does NOT alter extraction queries (single-number answers) or
-    queries that don't mention a key=value output pattern.
-    """
-    # Detect classification pattern: query mentions "key=N" format with
-    # multiple categories separated by spaces
-    _KV_PATTERN = re.compile(
-        r"\b(\w+=\s*[A-Z])\b.*\b(\w+=\s*[A-Z])\b",
-        re.IGNORECASE,
-    )
-    # More specific: looks for patterns like "positive=N negative=M neutral=K"
-    # or "category1=N category2=M" in the query's format instruction
-    _MULTI_KV_FORMAT = re.compile(
-        r"(\w+)=([A-Z_]\w*)\s+(\w+)=([A-Z_]\w*)",
-        re.IGNORECASE,
-    )
-    if not _MULTI_KV_FORMAT.search(query):
-        return query
-
-    # Extract the category names from the format pattern
-    format_match = _MULTI_KV_FORMAT.findall(query)
-    if not format_match:
-        return query
-
-    # Build format reinforcement suffix
-    suffix = (
-        "\n\nCRITICAL OUTPUT FORMAT: Your SUBMIT(answer=...) must contain ONLY "
-        "the counts in the exact format shown above (e.g. key1=N key2=M key3=K). "
-        "Do NOT include any explanation, prose, or extra text in the answer. "
-        "Do NOT wrap in quotes or add punctuation beyond the key=value pairs. "
-        "The answer string must match the pattern: word=number word=number ... "
-        "with single spaces between pairs."
-    )
-    return query + suffix
 
 
 def _resolve_delegate_context(
